Add files via upload
Airbnb, Inc. Stock Market Analysis 
Stock Market Analysis of Airbnb, Inc. (ABNB) from its Founding and Listing Years, which are 2020 to 2022
Airbnb, Inc., headquartered in San Francisco, California, is a short-term homestay and experience marketplace. The company works as a broker, charging a fee for each booking. Brian Chesky, Nathan Blecharczyk, and Joe Gebbia launched the company in 2008. Airbnb is an abbreviated form of the company's full name, AirBedandBreakfast.com.
https://www.kaggle.com/datasets/whenamancodes/airbnb-inc-stock-market-analysis